Policy: DocC on new public symbols, one type per file, file-scope responder probe
Aziz policy pass: document the new public composer symbols (draft store
witnesses, terminalInputText/selectedTerminalID, delegate defaults, the
composer accessory case, safeAreaInsetsDidChange); split ComposerDockIntent,
DelayingDraftStore, DraftSwitchOwnershipTests, ComposerDockProbeView, and
ComposerStoreProbe into their own files; replace the CurrentResponderProbe
namespace enum with a file-scope helper.
